make optimal algorithm use a circular buffer
With this the memory usage for large data sizes is a fraction of what it used to be, but the current implementation needs to be re-optimized for this layout
