+++ src/data/compression/match_length.rs
@@ -96,6 +96,8 @@ pub fn match_length(
     max_len: usize,
     skip: usize,
 ) -> usize {
+    debug_assert!(matched_pos < source);
+    
     match_length_simd(buffer, source + skip, matched_pos + skip, max_len - skip) + skip
 }
